refactor: useFetchData 생성
